editor/requirements.txt
fastapi==0.100.1
uvicorn[standard]==0.23.2
openai==1.10.0
instructor==0.4.8
tiktoken
python-docx
python-multipart
supabase
jwt
bcrypt
requests
numpy
matplotlib